test(e2e): retarget chat-mode session navigator to the ChatSidebar
The in-surface thread rail (.chat-thread-rail) is dropped in chat mode now
that the left ChatSidebar owns thread navigation, so the specs that drove
it there must target .chat-sidebar instead.

- rename chat-thread-rail.spec.ts → chat-sidebar-nav.spec.ts; assert the
  project folders ("Collapse/Expand <name> threads"), the searchbox, and
  the "No threads match your search" empty state on .chat-sidebar
- chat-task-chip-nav: open the task chat from .chat-sidebar

All three pass locally on the desktop project.
